# swagger.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "NeuroSuitUp API",
        "description": "Tools for communicating with the NeuroSuitUp database",
        "version": "1.0.0",
        "contact": {
        "name": "Apostolos Praftsiotis",
        "email": "apospraf@gmail.com"
        }
    },
    "servers": [
        {
            "url":"http://192.168.100.14:3000",
            "description":"Olpal server"
        },
        {
            "url":"https://locahost:3000",
            "description":"Temp localhost server"
        },
        {
            "url": "https://neurosuituptest.herokuapp.com/",
            "description": "Heroku free server"
        }
    ],
    "components": {
      "schemas": { 
...
        "LevelInsert": {
            "type": "object",
            "properties": {
                "codename": {
                    "type": "string",
                    "description": "The codename of the specific level"
                },
                "isPublic": {
                    "type": "boolean",
                    "description": "Whether or not the specific level is available to every instructor"
                },
                "exercises": {
                    "type": "array",
                    "description": "Array containing the exercises the define the specifc level",
                    "items":{
                        "type": "string",
                        "description": "The exercise id (ObjectId, mongodb)"
                    }
                },
                "delays": {
                    "type": "array",
                    "description": "The delay in seconds for each exercise pair",
                    "items": {
                        "type": "string",
                        "description": "The delay in seconds (sec)"
                    }
                },
                "belt": {
                    "type": "string",
                    "description": "The color of the belt (optional?)"
                }
            }
        },
        "Level": {
            "allOf": [
                {"$ref": "#/components/schemas/LevelInsert"},
                {
                    "type": "object",
                    "properties": {
                       "_id": {
                           "type": "string",
                           "description": "The exercise id (ObjectId, mongodb)"
                        }
                    }
                }
            ]
        },
        "ErrorMessage": {
            "type": "object",
            "properties": {
                "err": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        }, 
        "Message": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        }
      },
      "parameters": {
          "selectstring": {
              "in": "header",
              "name": "selectstring",
              "description": "Selects the fields of the returned resource",
              "schema": {
                  "type": "string"
              },
              "allowEmptyValue": true
          }
      },
      "responses": {
...
        "NotFound": {
            "description": "The specified resource was not found",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error" 
                    }
                }
            }
        }, 
        "Unauthorized": {
            "description": "Authorization failed, either by unsuccessfull verification of token, or by limited permissions",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error"
                    }
                }
            }
        }, 
        "InternalError": {
            "description": "Internal or structural error",
            "content": {
                "application/json": {
                    "schema": {
                        "$ref": "#/components/schemas/Error"
                    }
                }
            }
        }
      },
      "securitySchemes": {
        "bearerAuth": {
          "type": "http",
          "scheme": "bearer",
          "bearerFormat": "JWT"
        }
      }
    },
    "security": [
      {
        "bearerAuth": []
      }
    ],
    "paths": {
...
      "/levels/{levelId}": {
          "parameters": [{
            "in": "path",
            "name": "levelId",
            "required": true,
            "schema": {
                "type": "string"
            }
          }],
          "get": {
              "summary": "Get level by id",
              "tags": ["levels"],
              "parameters": [{
                    "$ref": "#/components/parameters/selectstring"
              }],
              "responses": {
                  "200":{
                      "description": "Query was successful, return found level",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Level"
                              }
                          }
                      }
                  },
                  "401": {
                      "$ref": "#/components/responses/Unauthorized"
                  },
                  "404": {
                      "$ref": "#/components/responses/NotFound"
                  }, 
                  "500": {
                      "$ref": "#/components/responses/InternalError"
                  }
              }
          },
          "delete": {
              "summary": "Delete exercise by id",
              "tags": ["levels"],
              "responses": {
                  "200": {
                      "description": "Successfully deleted level",
                      "content": {
                          "application/json":{
                              "schema":{
                                  "$ref": "#/components/schemas/Message"
                              }
                          }
                      }
                  },
                  "401": {
                      "$ref": "#/components/responses/Unauthorized"
                  },
                  "403": {
                      "description": "Level not owned",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/Message"
                              }
                          }
                      }
                  },
                  "404": {
                      "description": "User not found",
                      "content": {
                          "application/json": {
                              "schema": {
                                  "$ref": "#/components/schemas/ErrorMessage"
                              }
                          }
                      }
                  },
                  "500": {
                      "$ref": "#/components/responses/InternalError"
                  }
              }
          }
      },
      "/levels/": {
        "get": {
            "summary": "Get all available levels",
            "tags": ["levels"],
            "parameters": [{
                "$ref": "#/components/parameters/selectstring"
            }],
            "responses": {
                "200": {
                    "description": "Query was successful, return found levels",
                    "content": {
                        "application/json":{
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "count": {
                                        "type": "integer",
                                        "description": "The number of found levels"
                                    },
                                    "levels": {
                                        "type": "array",
                                        "items": {
                                            "$ref": "#/components/schemas/Level"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "401": {
                    "$ref": "#/components/responses/Unauthorized"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }                       
        },
        "post": {
            "summary": "Insert a new level to the database",
            "tags": ["levels"],
            "requestBody": {
                "description": "The new level",
                "required": true,
                "content": {
                    "application/x-www-form-urlencoded": {
                        "schema": {
                            "$ref": "#/components/schemas/LevelInsert"
                        }
                    }
                }
            },
            "responses": {
                "201": {
                    "description": "Level successfully inserted to the database",
                    "content": {
                        "application/json":{
                            "schema":{
                                "type": "object",
                                "properties": {
                                    "_id": {
                                        "type": "string",
                                        "description": "The Mongo ObjectId the inserted level"
                                    }
                                }
                            }
                        }
                    }
                },
                "401":{
                    "$ref": "#/components/responses/Unauthorized"
                },
                "409": {
                    "description": "Level codename already exists",
                    "content": {
                        "application/json":{
                            "schema":{
                                "$ref": "#/components/schemas/Message"
                            }
                        }
                    }
                },
                "500":{
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
      "/fetchLevels": {
        "post": {
            "summary": "Fetch specific levels by ids, located at the body of the request",
            "tags": ["levels"],
            "parameters": [
                {
                    "$ref": "#/components/parameters/selectstring"
                }
            ],
            "requestBody": {
                "description": "The array of the requested level ids",
                "content": {
                    "application/json": {
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string",
                                "description": "The level id (ObjectId mongo)"
                            }
                        },
                        "example": "{['615c94503b3ffd5cccec496a', '615c945e3b3ffd5cccec496b']}"
                    }
                }
            },
            "responses": {
                "200": {
                    "description": "Successfully fetched all requested levels",
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "count": {
                                        "type": "integer",
                                        "description": "The number of levels"
                                    },
                                    "tasks": {
                                        "type": "array",
                                        "description": "Array containing the level ids",
                                        "items": {
                                            "type": "string"
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "401":{
                    "$ref": "#/components/responses/Unauthorized"
                },
                "500": {
                    "$ref": "#/components/responses/InternalError"
                }
            }
        }
      },
...
    }
}